rename configurator components e.g. StagingAreaConfigFixture
