+++ components/src/hardware-sim/DeckConfigurator/index.tsx
@@ -13,10 +13,10 @@ import { COLORS } from '../../ui-style-constants'
 import { DeckSlotLocation } from '../DeckSlotLocation'
 import { SlotLabels } from '../Deck'
 import { RobotCoordinateSpace } from '../RobotCoordinateSpace'
-import { EmptyFixture } from './EmptyFixture'
-import { StagingAreaFixture } from './StagingAreaFixture'
-import { TrashBinFixture } from './TrashBinFixture'
-import { WasteChuteFixture } from './WasteChuteFixture'
+import { EmptyConfigFixture } from './EmptyConfigFixture'
+import { StagingAreaConfigFixture } from './StagingAreaConfigFixture'
+import { TrashBinConfigFixture } from './TrashBinConfigFixture'
+import { WasteChuteConfigFixture } from './WasteChuteConfigFixture'
 
 import type { Cutout, DeckConfiguration } from '@opentrons/shared-data'
 
@@ -85,28 +85,28 @@ export function DeckConfigurator(props: DeckConfiguratorProps): JSX.Element {
         </React.Fragment>
       ))}
       {stagingAreaFixtures.map(fixture => (
-        <StagingAreaFixture
+        <StagingAreaConfigFixture
           key={fixture.fixtureId}
           handleClickRemove={handleClickRemove}
           fixtureLocation={fixture.fixtureLocation}
         />
       ))}
       {emptyFixtures.map(fixture => (
-        <EmptyFixture
+        <EmptyConfigFixture
           key={fixture.fixtureId}
           handleClickAdd={handleClickAdd}
           fixtureLocation={fixture.fixtureLocation}
         />
       ))}
       {wasteChuteFixtures.map(fixture => (
-        <WasteChuteFixture
+        <WasteChuteConfigFixture
           key={fixture.fixtureId}
           handleClickRemove={handleClickRemove}
           fixtureLocation={fixture.fixtureLocation}
         />
       ))}
       {trashBinFixtures.map(fixture => (
-        <TrashBinFixture
+        <TrashBinConfigFixture
           key={fixture.fixtureId}
           handleClickRemove={handleClickRemove}
           fixtureLocation={fixture.fixtureLocation}
